fw: drivers: i2s: reorganize i2s init
description: moved both i2s inits to their own source and header files
