+++ firmware/Inc/io/interfaces/i2s/i2s_microphone.h
@@ -18,7 +18,10 @@
 #define I2S_MICROPHONE_PLLI2SR         2
 #define I2S_MICROPHONE_PERIPH_CLOCK    RCC_PERIPHCLK_I2S
 
+extern I2S_HandleTypeDef hi2s2;
+
 void i2s_microphone_init(void);
-void i2s_setup_receive_dma(void);
+void i2s2_init(void);
+bool i2s2_receive_dma(I2S_HandleTypeDef *hi2s, uint16_t *buff);
 
 #endif /* I2S_MICROPHONE_H_ */
