+++ firmware/Src/Drivers/io/interfaces/i2s/i2s_microphone.c
@@ -7,146 +7,49 @@
 
 #include "i2s_microphone.h"
 
-static bool i2s2_receive_dma(I2S_HandleTypeDef *hi2s, uint16_t *buff);
-
-DMA_HandleTypeDef hdma_spi2_rx;
-DMA_HandleTypeDef hdma_spi3_tx;
+I2S_HandleTypeDef hi2s2;
 
 struct microphone microphone;
 
-//void i2s_setup_receive_dma(void)
-//{
-//    i2s2_receive_dma(&hi2s2, &microphone.buff[0]);
-//
-//    if (!) {
-//        microphone.state == MICROPHONE_INIT_ERROR;
-//    }
-//}
-
-void HAL_I2S_MspInit(I2S_HandleTypeDef* i2sHandle)
+void i2s2_init(void)
 {
-    GPIO_InitTypeDef GPIO_InitStruct = {0};
+    hi2s2.Instance = SPI2;
+    hi2s2.Init.Mode = I2S_MODE_SLAVE_RX;
+    hi2s2.Init.Standard = I2S_STANDARD_MSB;
+    hi2s2.Init.DataFormat = I2S_DATAFORMAT_24B;
+    hi2s2.Init.MCLKOutput = I2S_MCLKOUTPUT_DISABLE;
+    hi2s2.Init.AudioFreq = I2S_AUDIOFREQ_48K;
+    hi2s2.Init.CPOL = I2S_CPOL_LOW;
+    hi2s2.Init.ClockSource = I2S_CLOCK_PLL;
+    hi2s2.Init.FullDuplexMode = I2S_FULLDUPLEXMODE_DISABLE;
 
-    if (i2sHandle->Instance==SPI2) {
-      __HAL_RCC_SPI2_CLK_ENABLE();
-
-      __HAL_RCC_GPIOC_CLK_ENABLE();
-      __HAL_RCC_GPIOB_CLK_ENABLE();
-      /**I2S2 GPIO Configuration
-      PC3     ------> I2S2_SD
-      PB10     ------> I2S2_CK
-      PB12     ------> I2S2_WS
-      */
-      GPIO_InitStruct.Pin = GPIO_PIN_3;
-      GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-      GPIO_InitStruct.Pull = GPIO_NOPULL;
-      GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-      GPIO_InitStruct.Alternate = GPIO_AF5_SPI2;
-      HAL_GPIO_Init(GPIOC, &GPIO_InitStruct);
-
-      GPIO_InitStruct.Pin = GPIO_PIN_10|GPIO_PIN_12;
-      GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-      GPIO_InitStruct.Pull = GPIO_NOPULL;
-      GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-      GPIO_InitStruct.Alternate = GPIO_AF5_SPI2;
-      HAL_GPIO_Init(GPIOB, &GPIO_InitStruct);
-
-      hdma_spi2_rx.Instance = DMA1_Stream3;
-      hdma_spi2_rx.Init.Channel = DMA_CHANNEL_0;
-      hdma_spi2_rx.Init.Direction = DMA_PERIPH_TO_MEMORY;
-      hdma_spi2_rx.Init.PeriphInc = DMA_PINC_DISABLE;
-      hdma_spi2_rx.Init.MemInc = DMA_MINC_ENABLE;
-      hdma_spi2_rx.Init.PeriphDataAlignment = DMA_PDATAALIGN_HALFWORD;
-      hdma_spi2_rx.Init.MemDataAlignment = DMA_MDATAALIGN_HALFWORD;
-      hdma_spi2_rx.Init.Mode = DMA_CIRCULAR;
-      hdma_spi2_rx.Init.Priority = DMA_PRIORITY_LOW;
-      hdma_spi2_rx.Init.FIFOMode = DMA_FIFOMODE_DISABLE;
-      if (HAL_DMA_Init(&hdma_spi2_rx) != HAL_OK) {
-        Error_Handler();
-      }
-
-      __HAL_LINKDMA(i2sHandle,hdmarx,hdma_spi2_rx);
-
-    }
-    else if(i2sHandle->Instance==SPI3) {
-      __HAL_RCC_SPI3_CLK_ENABLE();
-
-      __HAL_RCC_GPIOA_CLK_ENABLE();
-      __HAL_RCC_GPIOC_CLK_ENABLE();
-      __HAL_RCC_GPIOB_CLK_ENABLE();
-      /**I2S3 GPIO Configuration
-      PA4     ------> I2S3_WS
-      PC7     ------> I2S3_MCK
-      PC12    ------> I2S3_SD
-      PB3     ------> I2S3_CK
-      */
-      GPIO_InitStruct.Pin = GPIO_PIN_4;
-      GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-      GPIO_InitStruct.Pull = GPIO_NOPULL;
-      GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-      GPIO_InitStruct.Alternate = GPIO_AF6_SPI3;
-      HAL_GPIO_Init(GPIOA, &GPIO_InitStruct);
-
-      GPIO_InitStruct.Pin = GPIO_PIN_12|GPIO_PIN_7;
-      GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-      GPIO_InitStruct.Pull = GPIO_NOPULL;
-      GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-      GPIO_InitStruct.Alternate = GPIO_AF6_SPI3;
-      HAL_GPIO_Init(GPIOC, &GPIO_InitStruct);
-
-      GPIO_InitStruct.Pin = GPIO_PIN_3;
-      GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-      GPIO_InitStruct.Pull = GPIO_NOPULL;
-      GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-      GPIO_InitStruct.Alternate = GPIO_AF6_SPI3;
-      HAL_GPIO_Init(GPIOB, &GPIO_InitStruct);
-
-      hdma_spi3_tx.Instance = DMA1_Stream5;
-      hdma_spi3_tx.Init.Channel = DMA_CHANNEL_0;
-      hdma_spi3_tx.Init.Direction = DMA_MEMORY_TO_PERIPH;
-      hdma_spi3_tx.Init.PeriphInc = DMA_PINC_DISABLE;
-      hdma_spi3_tx.Init.MemInc = DMA_MINC_ENABLE;
-      hdma_spi3_tx.Init.PeriphDataAlignment = DMA_PDATAALIGN_HALFWORD;
-      hdma_spi3_tx.Init.MemDataAlignment = DMA_MDATAALIGN_HALFWORD;
-      hdma_spi3_tx.Init.Mode = DMA_CIRCULAR;
-      hdma_spi3_tx.Init.Priority = DMA_PRIORITY_LOW;
-      hdma_spi3_tx.Init.FIFOMode = DMA_FIFOMODE_DISABLE;
-
-      if (HAL_DMA_Init(&hdma_spi3_tx) != HAL_OK)
-      {
-        Error_Handler();
-      }
-
-      __HAL_LINKDMA(i2sHandle,hdmatx,hdma_spi3_tx);
+    if (HAL_I2S_Init(&hi2s2) != HAL_OK) {
+      Error_Handler();
     }
 }
 
-//bool i2s2_receive_dma(I2S_HandleTypeDef *hi2s, uint16_t *buff)
-//{
-//    return HAL_I2S_Receive_DMA(hi2s, buff, 64) == HAL_OK;
-//}
+bool i2s2_receive_dma(I2S_HandleTypeDef *hi2s, uint16_t *buff)
+{
+    return HAL_I2S_Receive_DMA(hi2s, buff, 64) == HAL_OK;
+}
 
 void HAL_I2S_RxHalfCpltCallback(I2S_HandleTypeDef *hi2s)
 {
-    indication_led();
     microphone.state = MICROPHONE_RX_STATE_1;
 }
 
 void HAL_I2S_RxCpltCallback(I2S_HandleTypeDef *hi2s)
 {
-    indication_led();
     microphone.state = MICROPHONE_RX_STATE_2;
 }
 
 void HAL_I2S_TxHalfCpltCallback(I2S_HandleTypeDef *hi2s)
 {
-    indication_led();
     microphone.state = MICROPHONE_TX_STATE_1;
 }
 
 void HAL_I2S_TxCpltCallback(I2S_HandleTypeDef *hi2s)
 {
-    indication_led();
     microphone.state = MICROPHONE_TX_STATE_2;
 }
